@@ -1,15 +1,24 @@
 apply plugin: 'com.android.library'
+apply plugin: 'maven-publish'
 
 android {
-    compileSdkVersion 28
+    compileSdk 34
+    namespace "segmented_control.widget"
+
+    publishing {
+        singleVariant("release") {
+            withSourcesJar()
+            withJavadocJar()
+        }
+    }
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 28
+        minSdkVersion 22
+        targetSdkVersion 33
         versionCode 1
         versionName "1.2.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
     }
     buildTypes {
@@ -22,16 +31,28 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    api 'com.github.RobertApikyan:ViewComponent:1.0.3'
-    api 'com.github.RobertApikyan:SecLayout:1.0.3'
+    implementation 'androidx.annotation:annotation:1.7.0'
+    implementation 'androidx.core:core:1.10.1'
+    implementation 'com.google.android.material:material:1.4.0'
+
+    api 'com.github.loconspzoo:ViewComponent:1.0'
+    api 'com.github.loconspzoo:SecLayout:1.0'
 }
 
 task sourcesJar(type: Jar) {
     from android.sourceSets.main.java.srcDirs
-    classifier = 'sources'
+    archiveClassifier.set("sources")
 }
 
-artifacts {
-    archives sourcesJar
+afterEvaluate {
+    publishing {
+        publications {
+            release(MavenPublication) {
+                from components.release
+                groupId = 'com.github.loconspzoo.SegmentedControl'
+                artifactId = 'android-segmented_control'
+                version = '1.0'
+            }
+        }
+    }
 }
