@@ -1,10 +1,9 @@
 package segmented_control.widget.custom.android.com.segmentedcontrol.custom_segment;
 
-import android.support.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 
-import segmented_control.widget.custom.android.com.segmented_control.R;
+import androidx.annotation.NonNull;
 import segmented_control.widget.custom.android.com.segmentedcontrol.item_row_column.SegmentAdapter;
 
 /**
@@ -16,6 +15,6 @@ public class SegmentAdapterImpl extends SegmentAdapter<CharSequence, SegmentView
     @NonNull
     @Override
     protected SegmentViewHolderImpl onCreateViewHolder(@NonNull LayoutInflater layoutInflater, ViewGroup viewGroup, int i) {
-        return new SegmentViewHolderImpl(layoutInflater.inflate(R.layout.item_segment_impl, null));
+        return new SegmentViewHolderImpl(layoutInflater.inflate(segmented_control.widget.R.layout.item_segment_impl, null));
     }
 }
