@@ -4,9 +4,9 @@ import android.animation.ObjectAnimator;
 import android.animation.ValueAnimator;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.GradientDrawable;
-import android.support.annotation.IntRange;
 import android.view.View;
 
+import androidx.annotation.IntRange;
 import segmented_control.widget.custom.android.com.segmentedcontrol.item_row_column.SegmentBackgroundType;
 import segmented_control.widget.custom.android.com.segmentedcontrol.item_row_column.SegmentViewHolder;
 
